Create study data set by merging and  cleaning the dataset
Because the dataset is small, when I searched for duplicate ["Mouse ID", "Timepoint"] data entries, I could see that the Tumor Size and Metastatic Site for the duplicate data differed. If the study was done on 10,000 mice, and there was tons of duplicate ["Mouse ID", "Timepoint"] data, I would want to know if the duplicate data was a double entry. If all duplicate ["Mouse ID", "Timepoint"] data for a specific Mouse ID was identical, I could assume that data is good and was just entered multiple times, therefore including it in the study data and deleting the duplicates. But if the ["Mouse ID", "Timepoint"] duplicate's ["Tumor Volume (mm3)","Metastatic Sites"] differed anywhere in the entire dataset, that "Mouse ID" would need to be excluded from the study data.