Made an assumption that I should have verified (Fixed)
Made an assumption that all the mice in the study made it to "Timepoint" == 45. That was not the case and I am glad that I caught it. This is another variable that is not constant and will need to be normalized or the data that  does not have a "Timepoint" == 45 should be removed.